# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GearUp API",
    "version": "1.0.0",
    "description": "Sports and outdoor gear rental service backend API"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/rentals": {
      "post": {
        "tags": [
          "Rentals"
        ],
        "summary": "Create a rental order"
      },
      "get": {
        "tags": [
          "Rentals"
        ],
        "summary": "List current customer's rental orders (paginated)",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ]
      }
    },
...
    "/api/provider/orders": {
      "get": {
        "tags": [
          "Provider"
        ],
        "summary": "View incoming rental orders (paginated)",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ]
      }
    },
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List all users (paginated)",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "CUSTOMER",
                "PROVIDER",
                "ADMIN"
              ]
            }
          }
        ]
      }
    },
...
    "/api/admin/gear": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List all gear listings (paginated)",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ]
      }
    },
    "/api/admin/rentals": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List all rental orders (paginated)",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ]
      }
    }
  },
...
}